requirements.txt
streamlit>=1.32.0
pandas>=2.2.0
numpy>=1.26.0
plotly>=5.19.0
seaborn>=0.13.0
matplotlib>=3.8.0
biopython>=1.83
python-dotenv>=1.0.0
pydantic>=2.6.0
typing-extensions>=4.9.0
pytest>=8.0.0
pytest-cov>=4.1.0
black>=24.1.0
pylint>=3.0.0
mypy>=1.8.0